# perf(ext/headers): use regex.test instead of .exec