@marcosc90 2023-08-10
Correct benchmarks, since I was using an outdated build without the recent perf additions.

**main**
```
cpu: 13th Gen Intel(R) Core(TM) i9-13900H
runtime: deno 1.36.1 (x86_64-unknown-linux-gnu)

benchmark        time (avg)        iter/s             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
Headers.get     132.15 ns/iter   7,567,412.3 (126.84 ns … 142.25 ns)  133.9 ns 139.93 ns 141.56 ns
```