@mmastrac 2023-08-10
If you want to try out an experiment we never landed, some testing we did a few months back showed that bit-testing is the absolute fastest way to do this

I _believe_ that this is the code we used for the experiment:

```
let a = []; for (var i = 0; i < 128; i++) { a[i >> 4] |= (HTTP_TOKEN_CODE_POINT_RE.test(String.fromCharCode(i)) ? 1 << (i & 15) : 0) }
```

```
  function bittest(val) {
    const BITMASK = [
        0,     0, 27898,
      1023, 65534, 51199,
      65535, 22527
    ];
    for (var i = 0; i < val.length; i++) {
      const c = val.charCodeAt(i);
      if ((BITMASK[c >> 4] & (c & 15)) == 0) {
        return false;
      }
    }
    return true;
  }
```