@marcosc90 2023-08-10
I'll do a bit of testing with that @mmastrac 

---

Regarding this PR, I noticed that @bartlomieju changed `.test` with `.exec` and resulted in a performance gain.

https://github.com/denoland/deno/pull/19364

Someone want to bench this and post results? All my benches result in `.test` being much faster than `.exec`