@mmastrac 2023-08-10
As @lino-levan said on the other PR

"The V8 engine is a whimsical place, seemingly random changes can have surprising differences in performance."

Are you using `tools/build_bench.ts` BTW? We built that tool to help w/benchmarking noise from incremental LTO builds.