@marcosc90 2023-08-11
I'm going to close for now, and do some further benching when I have some extra time.