@bartlomieju 2023-08-11
@marcosc90 I think we land this PR as is, we can run internal benches tomorrow to see if there's visible improvement.