@marcosc90 2023-08-11
@bartlomieju sounds good to me. I'm also going through all regex being used in headers/response to see if they can be optimized a bit.