@bartlomieju 2023-08-11
> @bartlomieju sounds good to me. I'm also going through all regex being used in headers/response to see if they can be optimized a bit.

Great to hear that as these regexes are showing up in the flamegraphs we plot. If you want to take a deeper look into what's going on when benching HTTP server I highly recommend using [`samply`](https://github.com/mstange/samply) - you can generate a flamegraph easily by doing `samply record -r 20000 deno run -A server.js`