@mmastrac 2023-08-12
Merged this one as is -- thanks for all the perf fixes!